refactor(cli): sink slash-command registry into internal/cli/prompts (#383)
Move repl's private buildSlashRegistry/loadPromptPaths/promptTemplateSources
(plus the registerLiveCommands/registerPluginCommands/presetListing helpers
they depend on) into a new shared package internal/cli/prompts, exported as
BuildSlashRegistry/LoadPromptPaths/PromptTemplateSources (+ RegisterLive/
PluginCommands) so the REPL and the forthcoming TUI wire the same command
surface from one owner. repl now calls the shared package; the private copies
are deleted. Prompt-template tests move to the prompts package; repl tests
that also exercise REPL-only helpers stay put and call the exported API.

Pure refactor + rename: REPL behavior is unchanged.
