+++ internal/cli/repl/help_line_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/smallnest/pigo/internal/cli"
+	"github.com/smallnest/pigo/internal/cli/prompts"
 	"github.com/smallnest/pigo/internal/runtime"
 )
 
@@ -36,7 +37,7 @@ func TestFormatHelpLine(t *testing.T) {
 // prompt template with its argument-hint, description, and source tier.
 func TestHelpActionIncludesTemplateLabelAndTier(t *testing.T) {
 	reg := runtime.NewSlashRegistry()
-	registerLiveCommands(reg, &cli.LiveConfig{Model: "test", ProviderName: "test"})
+	prompts.RegisterLiveCommands(reg, &cli.LiveConfig{Model: "test", ProviderName: "test"})
 	reg.AddUser(runtime.SlashCommand{
 		Name:         "review",
 		ArgumentHint: "<PR-URL>",
